+++ app/Http/Controllers/ImportController.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Kami\Cocktail\Http\Controllers;
 
 use Throwable;
+use Symfony\Component\Yaml\Yaml;
 use Illuminate\Http\JsonResponse;
 use Kami\Cocktail\Scraper\Manager;
 use Kami\Cocktail\Services\ImportService;
@@ -25,7 +26,7 @@ class ImportController extends Controller
             try {
                 $scraper = Manager::scrape($source);
             } catch (Throwable $e) {
-                abort(404, $e->getMessage());
+                abort(400, $e->getMessage());
             }
 
             $dataToImport = $scraper->toArray();
@@ -33,10 +34,20 @@ class ImportController extends Controller
 
         if ($type === 'json') {
             if (!is_array($source)) {
-                $source = json_decode($source);
+                if (!$source = json_decode($source)) {
+                    abort(400, 'Unable to parse the JSON string');
+                }
             }
 
             $dataToImport = $source;
+        }
+
+        if ($type === 'yaml' || $type === 'yml') {
+            try {
+                $dataToImport = Yaml::parse($source);
+            } catch (Throwable) {
+                abort(400, sprintf('Unable to parse the YAML string'));
+            }
         }
 
         if ($save) {